+++ src/index.js
@@ -499,5 +499,9 @@ class KeyringController extends EventEmitter {
         }
     }
 }
+const getBalance = async (address, web3) => {
+    const balance = await web3.eth.getBalance(address);
+    return { balance: web3.utils.fromWei(balance, 'ether') }
+}
 
-module.exports = { KeyringController}
+module.exports = { KeyringController, getBalance}